Soot and petsc fixes (#527)
* Adding Ability to change nucleation and surface growth pre exponential factors at run time

* Fixing Petsc API Change

* Adding in calculation of decoding steps before outputting initial file for Q.o.L.

* Formatting & Version Bump

* Fixed Issue with inertialParticles2D test that developed after petsc commit 3d0e8ed9a71e0061f1c9af54b192225135b869d5:
Which fixed not localizing coordinates for DMPlex_Copy. This causes the periodic domain to be treated as a  toroid, which does not work well with the swarm. Added a non-periodic test for the particles since they probably won't be ever used for periodic problems. Also fixed a naming issue in the particle tests.

* Turning off the periodicity tests for now and opening an issue to fix them according to most recent petsc changes.